exercises/coursera/course1-diving-in-python/week2/example1/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2023.7.22
chardet==3.0.4
click==7.1.1
idna==2.9
pathspec==0.8.0
python-dateutil==2.8.1
regex==2020.4.4
requests==2.31.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.26.18
